services/gateway/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.35
psycopg2-binary>=2.9.9
pydantic>=2.9.0
pydantic-settings>=2.5.0
redis>=5.0.0
httpx>=0.27.0
PyJWT>=2.9.0
